zng_wgt_input::cmd

Function on_copy

pub fn on_copy(child: impl IntoUiNode, handler: Handler<CommandArgs>) -> UiNode

P On copy command.

Receives COPY_CMD command events scoped on the widget. You can use the CLIPBOARD service to send data to the clipboard.

Preview

You can preview this command event using on_pre_copy. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.

Enabled

You can use the can_copy context property to disable the command handle. When disabled the handler is not called and the command handle signals as disabled.