zng_wgt_input::cmd

Function on_new

pub fn on_new(
    child: impl IntoUiNode,
    handler: impl WidgetHandler<CommandArgs>,
) -> UiNode

P On new command.

Receives NEW_CMD command events scoped on the widget. The command parameter can be the new item type identifier.

Preview

You can preview this command event using on_pre_new. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.

Enabled

You can use the can_new context property to disable the command handle. When disabled the handler is not called and the command handle signals as disabled.