zng_wgt_input::cmd

Function on_open

pub fn on_open(
    child: impl IntoUiNode,
    handler: impl WidgetHandler<CommandArgs>,
) -> UiNode

P On open command.

Receives OPEN_CMD command events scoped on the widget. The command parameter can be a path to open, otherwise the path must be derived from context or the user prompted.

You can use WINDOWS.native_file_dialog to prompt the user for a file or folder path.

Preview

You can preview this command event using on_pre_open. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.

Enabled

You can use the can_open context property to disable the command handle. When disabled the handler is not called and the command handle signals as disabled.