zng_wgt_input::cmd

Function on_pre_settings

pub fn on_pre_settings(
    child: impl IntoUiNode,
    handler: Handler<CommandArgs>,
) -> UiNode

P On settings command.

Receives SETTINGS_CMD command events scoped on the widget.

Command

This property will subscribe to the SETTINGS_CMD command scoped on the widget. If set on the Window! root widget it will also subscribe to the command scoped on the window.

The command handle is enabled by default and can be disabled using the contextual property can_settings .

Route

This event property uses the preview route, that is, the handler is called before the children widget handlers and before the $pn_ident handlers.