Function zng_wgt_input::cmd::on_settings

pub fn on_settings(
    child: impl UiNode,
    handler: impl WidgetHandler<CommandArgs>,
) -> impl UiNode

P On settings command.

Receives SETTINGS_CMD command events scoped on the widget.

Preview

You can preview this command event using on_pre_settings. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.