zng_wgt_input

Module drag_drop

Drag&drop properties, event properties.

Functions

draggable

P If this widget can be dragged in a drag&drop operation.

is_drag_hovered

P If the dragging cursor is over the widget or a descendant and the widget is enabled.

is_drag_hovered_disabled

P If the dragging cursor is over the widget or a descendant and the widget is disabled.

is_dragging

P If the draggable widget is dragging.

on_drag_end

P Draggable widget stopped dragging.

on_drag_enter

P Dragging cursor entered the widget area and the widget is enabled.

on_drag_hovered

P Dragging cursor entered or exited the widget area and the widget is enabled.

on_drag_leave

P Dragging cursor exited the widget area and the widget is enabled.

on_drag_start

P Draggable widget started dragging.

on_drop

P Dragging cursor dropped data in the widget area and the widget is enabled.

on_pre_drag_end

P Preview on_drag_end event.

on_pre_drag_enter

P Preview on_drag_enter event.

on_pre_drag_hovered

P Preview on_drag_hovered event.

on_pre_drag_leave

P Preview on_drag_leave event.

on_pre_drag_start

P Preview on_drag_start event.

on_pre_drop

P Preview on_drop event.