zng_wgt_input

Function click_mode

pub fn click_mode(
    child: impl IntoUiNode,
    mode: impl IntoVar<Option<ClickMode>>,
) -> UiNode

P Defines how click events are generated for the widget.

Setting this to None will cause the widget to inherit the parent mode, or ClickMode::default if no parent sets the click mode.