Function zng_wgt_input::focus::focus_click_behavior

pub fn focus_click_behavior(
    child: impl UiNode,
    behavior: impl IntoVar<FocusClickBehavior>,
) -> impl UiNode

P Behavior of a widget when a click event is send to it or a descendant.

When a click event targets the widget or descendant the behavior closest to the target is applied, that is if Exit is set in a parent, but Ignore is set on the target than the click is ignored. This can be used to create a effects like a menu that closes on click for command items, but not for clicks in sub-menu items.

Note that this property does not subscribe to any event, it only observes events flowing trough.