zng_wgt_input

Module gesture

Gesture events and control, on_click, click_shortcut and more.

These events aggregate multiple lower-level events to represent a user interaction. Prefer using these events over the events directly tied to an input device.

Structs

ClickArgs

CLICK_EVENT arguments.

Enums

Mnemonic

Defines the mnemonic char key that clicks the widget when pressed and focus is within the parent mnemonic scope.

Statics

ACTIVE_MNEMONICS_VAR

Inside an active mnemonic_scope this context var is a read-only map of the selected char for each descendant of the scope.

Traits

MnemonicWidgetInfoExt

Extension methods for widget info about mnemonic metadata.

Functions

click_shortcut

P Keyboard shortcuts that focus and clicks this widget.

context_click_shortcut

P Keyboard shortcuts that focus and context clicks this widget.

get_mnemonic

P Gets the mnemonic mode enabled for this widget or ancestor.

get_mnemonic_char

P Get the active mnemonic shortcut char for this widget or ancestor.

mnemonic

P Defines the mnemonic char key that clicks the widget when pressed and focus is within the parent mnemonic scope.

mnemonic_scope

P Defines a mnemonic shortcut scope.

mnemonic_txt

P Defines the inner text of a mnemonic parent widget.

on_any_click

P On widget click from any source and of any click count and the widget is enabled.

on_any_double_click

P On widget double click from any source and the widget is enabled.

on_any_single_click

P On widget click from any source but excluding double/triple clicks and the widget is enabled.

on_any_triple_click

P On widget triple click from any source and the widget is enabled.

on_click

P On widget click with the primary button and any click count and the widget is enabled.

on_context_click

P On widget click with the secondary/context button and the widget is enabled.

on_disabled_click

P On widget click from any source and of any click count and the widget is disabled.

on_double_click

P On widget double click with the primary button and the widget is enabled.

on_pre_any_click

P On widget click from any source and of any click count and the widget is enabled.

on_pre_any_double_click

P On widget double click from any source and the widget is enabled.

on_pre_any_single_click

P On widget click from any source but excluding double/triple clicks and the widget is enabled.

on_pre_any_triple_click

P On widget triple click from any source and the widget is enabled.

on_pre_click

P On widget click with the primary button and any click count and the widget is enabled.

on_pre_context_click

P On widget click with the secondary/context button and the widget is enabled.

on_pre_disabled_click

P On widget click from any source and of any click count and the widget is disabled.

on_pre_double_click

P On widget double click with the primary button and the widget is enabled.

on_pre_single_click

P On widget click with the primary button, excluding double/triple clicks and the widget is enabled.

on_pre_triple_click

P On widget triple click with the primary button and the widget is enabled.

on_single_click

P On widget click with the primary button, excluding double/triple clicks and the widget is enabled.

on_triple_click

P On widget triple click with the primary button and the widget is enabled.