zng_wgt_input::gesture

Function mnemonic_txt

pub fn mnemonic_txt(child: impl IntoUiNode, txt: impl IntoVar<Txt>) -> UiNode

P Defines the inner text of a mnemonic parent widget.

Note that the Label! widget automatically sets this to its own txt, this property can override the inner text. The text is used when the widget or parent mnemonic is FromTxt or Auto.