zng_wgt_input::gesture
Function on_any_double_click

pub fn on_any_double_click(
    child: impl UiNode,
    handler: impl WidgetHandler<ClickArgs>,
) -> impl UiNode

P On widget double click from any source and the widget is enabled.

This raises for all possible sources of CLICK_EVENT, but only when the click count is two. Use on_double_click to handle only primary button clicks.

Preview

You can preview this event using on_pre_any_double_click. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.