zng_wgt_input::gesture

Function on_context_click

pub fn on_context_click(
    child: impl IntoUiNode,
    handler: Handler<ClickArgs>,
) -> UiNode

P On widget click with the secondary/context button and the widget is enabled.

This raises only if the click is context.

Preview

You can preview this event using on_pre_context_click. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.