Function zng_wgt_input::gesture::on_double_click

pub fn on_double_click(
    child: impl UiNode,
    handler: impl WidgetHandler<ClickArgs>,
) -> impl UiNode

P On widget double click with the primary button and the widget is enabled.

This raises only if the click is primary and the click count is two. Use on_any_double_click to handle double clicks from any button.

Preview

You can preview this event using on_pre_double_click. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.