Module zng_wgt_input::mouse

Mouse events, on_mouse_move, on_mouse_enter, on_mouse_down and more.

There events are low level and directly tied to a mouse device. Before using them review the gesture events, in particular the on_click event.

Functions

on_disabled_mouse_any_click

P Mouse clicked on the disabled widget with any button, including repeat clicks.

on_disabled_mouse_click

P Mouse clicked on the disabled widget with the primary button, including repeat clicks.

on_disabled_mouse_hovered

P Mouse entered or left the widget and descendant widgets area, the widget is disabled and cursor capture allows it.

on_disabled_mouse_input

P Mouse button pressed or release while the cursor is over the widget, the widget is disabled and no cursor capture blocks it.

on_disabled_mouse_wheel

P Mouse wheel scrolled while pointer is hovering widget and it is disabled.

on_mouse_any_click

P Mouse clicked on the widget with any button and including repeat clicks and it is enabled.

on_mouse_any_double_click

P Mouse double clicked on the widget with any button and it is enabled.

on_mouse_any_single_click

P Mouse clicked on the widget with any button but excluding repeat clicks and it is enabled.

on_mouse_any_triple_click

P Mouse triple clicked on the widget with any button and it is enabled.

on_mouse_click

P Mouse clicked on the widget with the primary button including repeat clicks and it is enabled.

on_mouse_double_click

P Mouse double clicked on the widget with the primary button and it is enabled.

on_mouse_down

P Mouse button pressed while the cursor is over the widget, the widget is enabled and cursor capture allows it.

on_mouse_enter

P Mouse is now over the widget or a descendant widget, the widget is enabled and cursor capture allows it.

on_mouse_hovered

P Mouse entered or left the widget and descendant widgets area, the widget is enabled and cursor capture allows it.

on_mouse_input

P Mouse button pressed or released while the cursor is over the widget, the widget is enabled and no cursor capture blocks it.

on_mouse_leave

P Mouse is no longer over the widget or any descendant widget, the widget is enabled and cursor capture allows it.

on_mouse_move

P Mouse cursor moved over the widget and cursor capture allows it.

on_mouse_scroll

P Mouse wheel scrolled while pointer is hovering the widget and the pressed keyboard modifiers allow a scroll operation and the widget is enabled.

on_mouse_single_click

P Mouse clicked on the widget with the primary button excluding repeat clicks and it is enabled.

on_mouse_triple_click

P Mouse triple clicked on the widget with the primary button and it is enabled.

on_mouse_up

P Mouse button released while the cursor if over the widget, the widget is enabled and cursor capture allows it.

on_mouse_wheel

P Mouse wheel scrolled while pointer is hovering widget and it is enabled.

on_mouse_zoom

P Mouse wheel scrolled while pointer is hovering the widget and the pressed keyboard modifiers allow a zoom operation and the widget is enabled.

on_pre_disabled_mouse_any_click

P Preview on_disabled_mouse_any_click event.

on_pre_disabled_mouse_click

P Preview on_disabled_mouse_click event.

on_pre_disabled_mouse_hovered

P Preview on_disabled_mouse_hovered event.

on_pre_disabled_mouse_input

P Preview on_disabled_mouse_input event.

on_pre_disabled_mouse_wheel

P Preview on_disabled_mouse_wheel event.

on_pre_mouse_any_click

P Preview on_mouse_any_click event.

on_pre_mouse_any_double_click

P Preview on_mouse_any_double_click event.

on_pre_mouse_any_single_click

P Preview on_mouse_any_single_click event.

on_pre_mouse_any_triple_click

P Preview on_mouse_any_triple_click event.

on_pre_mouse_click

P Preview on_mouse_click event.

on_pre_mouse_double_click

P Preview on_mouse_double_click event.

on_pre_mouse_down

P Preview on_mouse_down event.

on_pre_mouse_enter

P Preview on_mouse_enter event.

on_pre_mouse_hovered

P Preview on_mouse_hovered event.

on_pre_mouse_input

P Preview on_mouse_input event.

on_pre_mouse_leave

P Preview on_mouse_leave event.

on_pre_mouse_move

P Preview on_mouse_move event.

on_pre_mouse_scroll

P Preview on_mouse_scroll event.

on_pre_mouse_single_click

P Preview on_mouse_single_click event.

on_pre_mouse_triple_click

P Preview on_mouse_triple_click event.

on_pre_mouse_up

P Preview on_mouse_up event.

on_pre_mouse_wheel

P Preview on_mouse_wheel event.

on_pre_mouse_zoom

P Preview on_mouse_zoom event.