Function zng_wgt_input::mouse::on_mouse_wheel

pub fn on_mouse_wheel(
    child: impl UiNode,
    handler: impl WidgetHandler<MouseWheelArgs>
) -> impl UiNode

P Mouse wheel scrolled while pointer is hovering widget and it is enabled.

Preview

You can preview this event using on_pre_mouse_wheel. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.