zng_wgt_input::mouse

Function on_pre_mouse_move

pub fn on_pre_mouse_move(
    child: impl IntoUiNode,
    handler: Handler<MouseMoveArgs>,
) -> UiNode

P Preview on_mouse_move event.

Preview

Preview event properties call the handler before the main event property and before the widget content, if you stop the propagation of a preview event the main event handler is not called.

Async

You can use async event handlers with this property, note that only the code before the fist .await is preview, subsequent code runs in widget updates.