zng_wgt_input

Module pointer_capture

Mouse and touch capture properties.

Enums

CaptureMode

Mouse and touch capture mode.

Functions

capture_pointer

P Capture mouse and touch for the widget on press.

capture_pointer_on_init

P Capture mouse and touch for the widget on init.

on_got_pointer_capture

P Widget acquired mouse and touch capture.

on_lost_pointer_capture

P Widget lost mouse and touch capture.

on_pointer_capture_changed

P Widget acquired or lost mouse and touch capture.

on_pre_got_pointer_capture

P Preview on_got_pointer_capture event.

on_pre_lost_pointer_capture

P Preview on_lost_pointer_capture event.

on_pre_pointer_capture_changed

P Preview on_pointer_capture_changed event.