zng_wgt_input

Module touch

Touch events, on_touch_move, on_touch_tap, on_touch_start and more.

There events are low level and directly tied to touch inputs. Before using them review the gesture events, in particular the on_click event.

Functions

on_disabled_touch_input

P Touch contact started or ended over the widget, it is disabled and cursor capture allows it.

on_disabled_touch_long_press

P Single touch contact was made and held in place for a duration of time (default 500ms) on the widget and the widget is disabled.

on_disabled_touch_tap

P Touch tap on the widget and it is disabled.

on_pre_disabled_touch_input

P Preview on_disabled_touch_input event.

on_pre_disabled_touch_long_press

P Preview on_disabled_touch_long_press event.

on_pre_disabled_touch_tap

P Preview on_disabled_touch_tap event.

on_pre_touch_cancel

P Preview on_touch_cancel event.

on_pre_touch_end

P Preview on_touch_end event.

on_pre_touch_enter

P Preview on_touch_enter event.

on_pre_touch_input

P Preview on_touch_input event.

on_pre_touch_leave

P Preview on_touch_leave event.

on_pre_touch_long_press

P Preview on_touch_long_press event.

on_pre_touch_move

P Preview on_touch_move event.

on_pre_touch_start

P Preview on_touch_start event.

on_pre_touch_tap

P Preview on_touch_tap event.

on_pre_touch_transform

P Preview on_touch_transform event.

on_pre_touched

P Preview on_touched event.

on_touch_cancel

P Touch contact canceled over the widget, it is enabled and cursor capture allows it.

on_touch_end

P Touch contact ended over the widget, it is enabled and cursor capture allows it.

on_touch_enter

P Touch contact is now over the widget or a descendant and it is enabled.

on_touch_input

P Touch contact started or ended over the widget, it is enabled and cursor capture allows it.

on_touch_leave

P Touch contact is no longer over the widget or any descendant and it is enabled.

on_touch_long_press

P Single touch contact was made and held in place for a duration of time (default 500ms) on the widget and the widget is enabled.

on_touch_move

P Touch contact moved over the widget and cursor capture allows it.

on_touch_start

P Touch contact started over the widget, it is enabled and cursor capture allows it.

on_touch_tap

P Touch tap on the widget and it is enabled.

on_touch_transform

P Touch gesture to translate, scale or rotate happened over this widget.

on_touched

P Touch contact entered or left the widget and descendants area and it is enabled.