zng_wgt_input::touch

Function on_pre_touch_enter

pub fn on_pre_touch_enter(
    child: impl IntoUiNode,
    handler: impl WidgetHandler<TouchedArgs>,
) -> UiNode

P Preview on_touch_enter event.

Preview

Preview event properties call the handler before the main event property and before the widget content, if you stop the propagation of a preview event the main event handler is not called.

Async

You can use async event handlers with this property, note that only the code before the fist .await is preview, subsequent code runs in widget updates.