zng_wgt_input::touch

Function on_touch_tap

pub fn on_touch_tap(
    child: impl IntoUiNode,
    handler: impl WidgetHandler<TouchTapArgs>,
) -> UiNode

P Touch tap on the widget and it is enabled.

Preview

You can preview this event using on_pre_touch_tap. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.