Crate zng_wgt_input

Input events and focus properties.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 1 feature flag, not enabled by default.

"drag_drop"

Enable drag&drop.

Modules

cmd

Common commands.

drag_drop

Drag&drop properties, event properties.

focus

Keyboard focus properties, tab_index, focusable, on_focus, is_focused and more.

gesture

Gesture events and control, on_click, click_shortcut and more.

keyboard

Keyboard events, on_key_down, on_key_up and more.

mouse

Mouse events, on_mouse_move, on_mouse_enter, on_mouse_down and more.

pointer_capture

Mouse and touch capture properties.

touch

Touch events, on_touch_move, on_touch_tap, on_touch_start and more.

Structs

CursorImg

Window custom cursor.

Enums

CursorIcon

Named system dependent cursor icon.

CursorSource

Window cursor source.

Functions

click_mode

P Defines how click events are generated for the widget.

cursor

P Sets the mouse pointer cursor displayed when hovering the widget.

is_cap_hovered

P If the mouse pointer is over the widget, or a descendant, or is captured by the it.

is_cap_mouse_pressed

P If the mouse pointer is pressed or captured by the widget and it is enabled.

is_cap_pointer_pressed

P If is_cap_mouse_pressed or is_cap_touched_from_start.

is_cap_pressed

P If is_cap_mouse_pressed, is_cap_touched_from_start or is_shortcut_pressed.

is_cap_touched

P If a touch contact point is over the widget, or is over a descendant, or is captured by it.

is_cap_touched_from_start

P If a touch contact point is over the widget, or is over a descendant, or is captured by it.

is_hovered

P If the mouse pointer is over the widget or a descendant and the widget is enabled.

is_hovered_disabled

P If the mouse pointer is over the widget or a descendant and the widget is disabled.

is_mouse_pressed

P If the mouse pointer is pressed in the widget and it is enabled.

is_pointer_pressed

P If is_mouse_pressed or is_touched_from_start.

is_pressed

P If is_mouse_pressed, is_touched_from_start or is_shortcut_pressed.

is_shortcut_pressed

P If the widget was clicked by shortcut or accessibility event and the shortcut_pressed_duration has not elapsed.

is_touched

P If a touch contact point is over the widget or a descendant and the it is enabled.

is_touched_from_start

P If a touch contact that started over the widget is over it and it is enabled.

touch_transform

P Applies transforms from touch gestures on the widget.