zng_wgt_input

Function cursor

pub fn cursor(
    child: impl IntoUiNode,
    cursor: impl IntoVar<CursorSource>,
) -> UiNode

P Sets the mouse pointer cursor displayed when hovering the widget.

You can set this property to a CursorIcon for a named platform dependent icon, CursorImg for a custom image, or to false that converts to CursorSource::Hidden.