zng_wgt_input

Module focus

Keyboard focus properties, tab_index, focusable, on_focus, is_focused and more.

Structs

FocusableMix

m Focusable widget mixin. Enables keyboard focusing on the widget and adds a focused highlight visual.

Enums

FocusClickBehavior

Behavior of a widget when a click event is send to it or a descendant.

Statics

FOCUS_HIGHLIGHT_OFFSETS_VAR

Padding offsets of the foreground highlight when the widget is focused.

FOCUS_HIGHLIGHT_SIDES_VAR

Border sides of the foreground highlight when the widget is focused.

FOCUS_HIGHLIGHT_WIDTHS_VAR

Border widths of the foreground highlight when the widget is focused.

Functions

alt_focus_scope

P Widget is the ALT focus scope.

directional_nav

P Keyboard arrows navigation within this focus scope.

focus_click_behavior

P Behavior of a widget when a click event is send to it or a descendant.

focus_highlight

P Sets the foreground highlight values used when the widget is focused and highlighted.

focus_on_init

P If the widget is focused on info init.

focus_scope

P Makes the widget into a focus scope when set to true.

focus_scope_behavior

P Behavior of a focus scope when it receives direct focus.

focus_shortcut

P Keyboard shortcuts that focus this widget or its first focusable descendant or its first focusable parent.

focusable

P Makes the widget focusable when set to true.

is_focus_within

P If the widget or one of its descendants has keyboard focus.

is_focus_within_hgl

P If the widget or one of its descendants has keyboard focus and the user is using the keyboard to navigate.

is_focused

P If the widget has keyboard focus.

is_focused_hgl

P If the widget has keyboard focus and the user is using the keyboard to navigate.

is_return_focus

P If the widget will be focused when a parent scope is focused.

is_return_focus_within

P If the widget or one of its descendants will be focused when a focus scope is focused.

on_blur

P Widget lost direct keyboard focus.

on_focus

P Widget got direct keyboard focus.

on_focus_changed

P Focus changed in the widget or its descendants.

on_focus_enter

P Widget or one of its descendants got focus.

on_focus_leave

P Widget or one of its descendants lost focus.

on_pre_blur

P Preview on_blur event.

on_pre_focus

P Preview on_focus event.

on_pre_focus_changed

P Preview on_focus_changed event.

on_pre_focus_enter

P Preview on_focus_enter event.

on_pre_focus_leave

P Preview on_focus_leave event.

return_focus_on_deinit

P If the widget return focus to the previous focus when it inited.

skip_directional

P If directional navigation from outside this widget skips over it and its descendants.

tab_index

P Customizes the widget order during TAB navigation.

tab_nav

P Tab navigation within this focus scope.