zng_wgt_input

Module gesture

Gesture events and control, on_click, click_shortcut and more.

These events aggregate multiple lower-level events to represent a user interaction. Prefer using these events over the events directly tied to an input device.

Structs

ClickArgs

CLICK_EVENT arguments.

Functions

click_shortcut

P Keyboard shortcuts that focus and clicks this widget.

context_click_shortcut

P Keyboard shortcuts that focus and context clicks this widget.

on_any_click

P On widget click from any source and of any click count and the widget is enabled.

on_any_double_click

P On widget double click from any source and the widget is enabled.

on_any_single_click

P On widget click from any source but excluding double/triple clicks and the widget is enabled.

on_any_triple_click

P On widget triple click from any source and the widget is enabled.

on_click

P On widget click with the primary button and any click count and the widget is enabled.

on_context_click

P On widget click with the secondary/context button and the widget is enabled.

on_disabled_click

P On widget click from any source and of any click count and the widget is disabled.

on_double_click

P On widget double click with the primary button and the widget is enabled.

on_pre_any_click

P Preview on_any_click event.

on_pre_any_double_click

P Preview on_any_double_click event.

on_pre_any_single_click

P Preview on_any_single_click event.

on_pre_any_triple_click

P Preview on_any_triple_click event.

on_pre_click

P Preview on_click event.

on_pre_context_click

P Preview on_context_click event.

on_pre_disabled_click

P Preview on_disabled_click event.

on_pre_double_click

P Preview on_double_click event.

on_pre_single_click

P Preview on_single_click event.

on_pre_triple_click

P Preview on_triple_click event.

on_single_click

P On widget click with the primary button, excluding double/triple clicks and the widget is enabled.

on_triple_click

P On widget triple click with the primary button and the widget is enabled.