zng_wgt_input

Module keyboard

Keyboard events, on_key_down, on_key_up and more.

These events are low level and directly tied to a keyboard device. Before using them review the gesture properties, in particular the click_shortcut property.

Functions

on_disabled_key_input

P Event fired when a keyboard key is pressed or released and the widget is disabled.

on_key_down

P Event fired when a keyboard key is pressed and the widget is enabled.

on_key_input

P Event fired when a keyboard key is pressed or released and the widget is enabled.

on_key_up

P Event fired when a keyboard key is released and the widget is enabled.

on_pre_disabled_key_input

P Preview on_disabled_key_input event.

on_pre_key_down

P Preview on_key_down event.

on_pre_key_input

P Preview on_key_input event.

on_pre_key_up

P Preview on_key_up event.